feat: Add authentication
Authenticates against the ETRM API and stores bearer token and user data
in a cookie.
This cookie is never sent to the backend. The bearer token is taken from
the cookie and appended to the headers whenever a request is made.
Stored used data will be useful for role-based authorization.

Tested manually.