feat: make inline links pop out more
With "inline links" I mean those that are not part of buttons.
Making links a different color to normal text is an affordance. It makes
clear that this text can be interacted with.

Defined a secondary color for the app palette and applied it to text in
select links.

This secondary color will be useful later for interactions with icons.