feat: add change password form
- Add toaster and toast functionality
useToast() hook allows us to summon toasts from components.

- Create fetchWithToken action
this server action allows us to call fetchWithToken() from client
components (such as forms!). Because of the way server actions work,
they cannot simply return a promise like fetchWithToken() does.

instead, fetchWithToken action returns the already deserialized object.
with status, whether or not the response is ok, and optionally a body,
as represented in the ClientResponse interface.

This is pretty sparse but so far is what we need. We can always extend
the interface later.